Movement, Boxes, Collisions, and Splash Screens
The biggest complaint about HtR I used to get was that it was awful*.

The second biggest complaint was that it was disconcerting that the game simply restarted
when you died, with no indication of what had happened.

So, I present to you <drumroll> *Splash* Screens! With glorious SVGs!
The SVGs may need some resizing/recolouring work. But I doubt that work will happen.

The splash screens do tend to occasionally cause the whole browser to crash,
but that should only be a minor inconvenience, as no one should be using their browser
for anything else anyway.

I did some other stuff, but that pales in comparison to the glory of the Splash screens.

Next: Foxes! Music! Profit? Mwuhaha

* Although some people did actually say it was awful in a good way...

Gah, why is my EDITOR nano