Add theme Music
Tried to get MIDI.js to work, but it is broken.

Using HTML5 audio element and .ogg file seems to work.

Now that I have an amazing Splash Screen, I think I need to make a death sound effect to go with it.
Otherwise the player is not constantly being assaulted by noise, and that is a crying shame.

No victory condition yet... (traditionally it was when the music stopped).