Add an attack against Foxes (!)
It's just not satisfying that the only way to beat the pesky foxes
is to lure them under a box. Also it's really hard...

So, we have another way.

Pressing the down arrow whilst airborn will make you fall faster.
It adds a whole new dimension (downwards) to the game.

If you land on the *back* of a fox whilst going fast enough you will put her to sleep.
This is still really hard to do. I play tested for 15 minutes and only got it to happen twice...

When the fox is asleep it can be pushed like a box. But if you push her too hard
she will wake up and after a short while be able to eat you again.

This kind of makes the game funner because its now possible to build towers
just so you can push the top onto a sleeping fox.

This thing is beginning to suffer from feature creep...