Add survival timer to increase dramatic tension
Also made a CGI script to upload a function to control the player.
Seems a bit pointless but whatever.

Also Fixed Fox Sleepy Jump bug.

Got to 1:48 ... probably a personal best.

I am kind of mixing deprecated HTML4 things in but screw CSS.

Next: Will make it upload some stats so I can plot stuff.
This is based on the assumption that people will actually play it...

Seems a bit pointless, but whatever.