More Player information in view.py
List of all players.
Nice link to "you" in this list and also in the High Scores.

Other minor things: Last Player, Active Players

These are mainly so that I can easily check whether
people are still playing the game.

I'm adopting a "If people are still playing I will keep working on this" approach.