Data for Ox and Second Theme Song
My plan was to produce a new "level" with new music and some other element
every time someone beat the hardest level.

But I really need to make it harder, because this will buy me only a couple of minutes.
and it took me ~4 hours just to record the new music... and another 4 hours on and off to do the Ox :S

The new music is mostly based off something called "Humphrey The Rabbit" that I came up with ages before this game.
I can't remember why; it inspired the game, but for some reason I decided not to use it.

Next: Actually putting it in the game!