Add (better) support for smartphones + More tutorial
Test "devices" are
1) Samsung Galaxy Young (Android Gingerbread) + Mozilla Firefox
2) Firefox OS Simulator (!) running in Iceweasel

Use Javascript hacks (because the Right Way doesn't work)
to make the canvas take up the inner window instead of whatever the browser feels
like zooming it to. Also got rid of images and just detect touch events in the canvas.

The difficulty of the game becomes highly variable with device,
since the sprites and collision boxes do not scale but the velocities and
available space do. I don't think this is necessarily a bad thing.

Needed to make performance adjustments.
MainLoop chooses the next timeout based on the actual time between callbacks.

Refactored* the tutorial and added some more things.

* I'm not sure whether it is better or worse...