Enable external advertisements again
Yeah I'm an evil capitalist bastard and all that.

If you set the "adblock" cookie to something other than "yes"
you will hide the external advert. If you set it to "yes" it hides
the internal adverts (the ones that I wrote myself) as well.

You should also be able to resize the screen to hide it.

External advert is not under git for obvious reasons.