Replace CGI adverts.py with the index.json, and fix advertisements
The adblock "meta" joke actually does something when adblock is enabled!
Only took me 5 years to ~work out~ (google search) the code

Other improvements and refactoring where needed
Really the whole splash screen and texture loading needs Promises
But can't be fucked rewriting the whole thing

It works even if it is hacky and prone to race conditions between
the Game() and the first ChooseAdvert() calls.

Noticed a bug where if img.src is a 404, the SplashScreen hangs forever

PS:
Regarding d0586db1e8be905ecb204b7a75f05497ffe54798

sulix I removed your ad because you're not a dummy,
and I don't know if you'd still want an image that insults you
to be on public websites (even if it only gets 1 visit a year by GoogleBot).

I removed the UCC reference because it's not hosted on UCC
and I'd rather not think about the UCC.

I kept the Mylah reference though because it's true that it's no
longer hosted on mylah, and I think it is still funny.