Fix for AddEnemy timeout pausing all other timeouts
The problem is that the "paused" state is both a global game state,
and a state for individual timeouts.

What happens is that code for changing the level, will set the paused state,
but then create a timer anyway. That timer is ignoring the global game paused state.

However, other timers check the global game state,
 and they can potentially "as a workaround" pause the timers that
were created after the game was paused. That creates the race condition.

Perhaps browsers just got better at running setTimeout actually on time,
or maybe this started happening after some changes to the spawn rate,
but it just happens that the AddEnemy timeout coincides with a level change,
and obviously it didn't before when the code was "working".

Maybe a "safer" way to do this is to incorporate a timestamp into the
paused state, rather than simply having a boolean.

Anyway, tl;dr: "Game Paused != Timers Paused" except when it does.